Input/output

FiniteStateTransducers.txt2fst — Function

txt2fst(path,isym[,osym];W=TropicalWeight{Float32})

Loads a WFST from a text file using OpenFST format.

isym and osym can be either dictionaries or paths to the input/output symbol list.

FiniteStateTransducers.txt2sym — Function

txt2sym(path)

Loads a symbol list from a text file into a Dict using OpenFST convention.

FiniteStateTransducers.fst2dot — Function

fst2dot(fst)

Converts the fst to a string in the dot format.

Useful to visualise large FSTs using Graphviz.

FiniteStateTransducers.fst2pdf — Function

fst2pdf(fst,pdffile)

Prints the fst to a pdf file using Graphviz.

Requires Graphviz to be intalled.